requirements.txt
streamlit>=1.20
openai>=1.0.0
pydantic>=1.10
python-dotenv>=1.0
pytest>=7.0
requests>=2.28
